Add grade report, restructure code
The demo script now has the ability to generate a grade report,
summarizing information stored in eDnevnik. The report is stored
to "report.txt".

The code has been changed a bit, but the report function and the
normal output could definitely be unified in one, as there is a lot
of repeating code right now.