Preparation for Electron
Flask has been added as a dependency. It will be used to host an
HTTP API for the Electron frontend. The login part is so far
complete, but more need to follow.

The EDAP project's files have been moved into a separate folder.

Some skeleton files have been added for the Electron part.