Switch from dictionary to Redis
Also had to make a pretty bad decision here: storing passwords in 
plaintext. As CARNet does not provide an official API for eDnevnik, I 
cannot use a token to refresh the profile (the session expires quickly), 
so a reauthentication is required. I will try to find a better solution 
for this at some point.