Netrix: Switch from AdMobPro to AdMobFree [desc]
AdMobPro has been replaced with AdMobFree. This change is made for the
following reasons:
1. Closed-source code: impossible to know what it does behind the scenes
2. Fetching unauthorized remote domains: see http://bit.ly/2YZ0oFA.
This domain is apparently used to determine the sharing percentage.
Since this is an external server from an untrusted third party, it could
cause network problems and lost revenue. This is also not documented
anywhere on the plugin's page.
3. Lying about ad sharing rate, censoring discussion: see
http://bit.ly/2N3U0GR
Developer claims 2% of revenue is shared, code analysis says 30%. Users
create issue, call him out, he keeps deleting posts and does not state
any reason why he does so.