eDAP: Halve data processing time
By removing the `dumpable_logs` functionality and putting __edlog
behind a single `if` statement, it skips the whole inspect.stack() run.

This results in an about 46% faster profile build with 78% less function
calls being made in the process.

This commit also removes `dump_data` and the `return_processing_time`
functionality.